pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "diropt"
dynamic = ["version"]
description = "A library for distributionally robust learning in pytorch."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    {name = "Ronak Mehta", email = "ronakdm@uw.edu"},
]
keywords = [
    "python",
    "machine learning",
    "distribtuionally robust optimization"
]
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "absl-py>=0.7.1",
  "numpy>=1.21",
  "torch>=1.6",
  "numba>=0.58.1"
]

[project.optional-dependencies]
docs = [
    "sphinx>=6.0.0",
    "sphinx-book-theme>=1.0.1",  # Older versions fail to pin pydata-sphinx-theme
    "sphinxcontrib-katex",
    "sphinx-autodoc-typehints",
    "ipython>=8.8.0",  # 8.7.0 has ipython3 lexer error
    "myst-nb>=1.0.0",
    "matplotlib>=3.5.0",
    "sphinx-gallery>=0.14.0",
    "sphinx-collections>=0.0.1",
    "sphinx_contributors",
]

[tool.setuptools.packages.find]
include = ["README.md", "LICENSE"]
exclude = ["*_test.py"]
